refactor(auth): moved useVerifyOtp to useAuth to better reduce use of files
