+++ web/components/auth/otp-form.tsx
@@ -5,8 +5,7 @@ import { useSearchParams } from "next/navigation";
 import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { z } from "zod";
-
-import useVerifyOtp from "@/hooks/useVerifyOtp";
+import { useVerifyOtp } from "@/hooks/useAuth";
 
 const OtpSchema = z.object({
   otp: z.string().min(6, { message: "Invalid OTP" }),
